Fixed some more bugs
Fixed bug where application would crash if you removed a subject before logging anything to that subject

+++ app/src/main/java/com/example/lakethomason/bluetoothinteractiontest/PolarH7.java
@@ -124,9 +124,11 @@ public class PolarH7 {
     }
 
     public void stopLoggingAndDestroy() {
-        log = false;
-        fileCreator.closeFile();
-        fileCreator.deleteFile();
+        if (log) {
+            log = false;
+            fileCreator.closeFile();
+            fileCreator.deleteFile();
+        }
     }
 
     public File getFile() {
@@ -140,6 +142,9 @@ public class PolarH7 {
                 mHeartRateText.setText(String.valueOf(value));
                 if (value == 0) {
                     mHeartRateText.setVisibility(View.INVISIBLE);
+                }
+                else {
+                    updatePolarCheckBox(true);
                 }
             }
         });
